# backend/src/main/resources/openapi/tracex-api.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:62436
  description: Generated server url
paths:
...
  /api/v1/auth/verify-reset-otp:
    post:
      tags:
      - auth-controller
      operationId: verifyResetOtp
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyResetOtpDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseMapStringString'
  /api/v1/auth/verify-otp:
    post:
      tags:
      - auth-controller
      operationId: verifyOtp
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyOtpDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseLoginResponse'
  /api/v1/auth/verify-otp/resend:
    post:
      tags:
      - auth-controller
      operationId: resendOtp
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResendOtpDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseMapStringString'
  /api/v1/auth/reset-password:
    post:
      tags:
      - auth-controller
      operationId: resetPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseMapStringString'
  /api/v1/auth/requests/{id}/resend:
    post:
      tags:
      - access-request-controller
      operationId: resendInvite
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseAccessRequestSummaryDto'
  /api/v1/auth/requests/{id}/reject:
    post:
      tags:
      - access-request-controller
      operationId: rejectRequest
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RejectRequestDto'
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseAccessRequestSummaryDto'
  /api/v1/auth/requests/{id}/approve:
    post:
      tags:
      - access-request-controller
      operationId: approveRequest
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseAccessRequestSummaryDto'
  /api/v1/auth/request-access:
    post:
      tags:
      - auth-controller
      operationId: requestAccess
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestAccessDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseAccessRequestSummaryDto'
...
  /api/v1/auth/me/change-password:
    post:
      tags:
      - auth-controller
      operationId: changePassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseLoginResponse'
  /api/v1/auth/login:
    post:
      tags:
      - auth-controller
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseLoginResponse'
  /api/v1/auth/forgot-password:
    post:
      tags:
      - auth-controller
      operationId: forgotPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseMapStringString'
  /api/v1/auth/activate:
    post:
      tags:
      - auth-controller
      operationId: activate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivateAccountDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseMapStringObject'
...
  /api/v1/auth/me:
    get:
      tags:
      - auth-controller
      operationId: getMe
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseUserSummaryDto'
    patch:
      tags:
      - auth-controller
      operationId: updateMe
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseUserSummaryDto'
...
  /api/v1/auth/requests/{id}:
    delete:
      tags:
      - access-request-controller
      operationId: deleteRequest
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
components:
  schemas:
...
    ApiResponseVoid:
      type: object
      properties:
        success:
          type: boolean
        requestId:
          type: string
        data:
          type: object
        message:
          type: string
        code:
          type: string
        error:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldErrorDto'
    FieldErrorDto:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
...
    VerifyResetOtpDto:
      required:
      - email
      - otp
      type: object
      properties:
        email:
          type: string
        otp:
          pattern: "^\\d{6}$"
          type: string
    ApiResponseMapStringString:
      type: object
      properties:
        success:
          type: boolean
        requestId:
          type: string
        data:
          type: object
          additionalProperties:
            type: string
        message:
          type: string
        code:
          type: string
        error:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldErrorDto'
    VerifyOtpDto:
      required:
      - email
      - otp
      type: object
      properties:
        email:
          type: string
        otp:
          pattern: "^\\d{6}$"
          type: string
    ApiResponseLoginResponse:
      type: object
      properties:
        success:
          type: boolean
        requestId:
          type: string
        data:
          $ref: '#/components/schemas/LoginResponse'
        message:
          type: string
        code:
          type: string
        error:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldErrorDto'
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/UserSummaryDto'
    UserSummaryDto:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum:
          - admin
          - manager
          - factory-manager
          - quality-inspector
          - dispatch-coordinator
        emailVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        active:
          type: boolean
        superAdmin:
          type: boolean
        deleted:
          type: boolean
    ResendOtpDto:
      required:
      - email
      type: object
      properties:
        email:
          type: string
    ResetPasswordDto:
      required:
      - email
      - newPassword
      - resetToken
      type: object
      properties:
        email:
          type: string
        resetToken:
          type: string
        newPassword:
          maxLength: 2147483647
          minLength: 8
          type: string
    AccessRequestSummaryDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum:
          - admin
          - manager
          - factory-manager
          - quality-inspector
          - dispatch-coordinator
        status:
          type: string
          enum:
          - pending
          - approved
          - rejected
        note:
          type: string
        inviteExpiry:
          type: string
          format: date-time
        inviteUsed:
          type: boolean
        approvedBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ApiResponseAccessRequestSummaryDto:
      type: object
      properties:
        success:
          type: boolean
        requestId:
          type: string
        data:
          $ref: '#/components/schemas/AccessRequestSummaryDto'
        message:
          type: string
        code:
          type: string
        error:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldErrorDto'
    RejectRequestDto:
      type: object
      properties:
        note:
          type: string
    RequestAccessDto:
      required:
      - email
      - name
      - role
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
    ChangePasswordDto:
      required:
      - currentPassword
      - newPassword
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          maxLength: 2147483647
          minLength: 8
          type: string
    LoginRequest:
      required:
      - password
      - username
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    ForgotPasswordDto:
      required:
      - email
      type: object
      properties:
        email:
          type: string
    ActivateAccountDto:
      required:
      - password
      - token
      type: object
      properties:
        token:
          type: string
        password:
          maxLength: 2147483647
          minLength: 8
          type: string
    ApiResponseMapStringObject:
      type: object
      properties:
        success:
          type: boolean
        requestId:
          type: string
        data:
          type: object
          additionalProperties:
            type: object
        message:
          type: string
        code:
          type: string
        error:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldErrorDto'
...
    ApiResponseUserSummaryDto:
      type: object
      properties:
        success:
          type: boolean
        requestId:
          type: string
        data:
          $ref: '#/components/schemas/UserSummaryDto'
        message:
          type: string
        code:
          type: string
        error:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldErrorDto'
...
    UpdateProfileDto:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
...